git-svn-id: https://svn.code.sf.net/p/wired/code/trunk/wired@1377 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ src/gui/SettingWindow.h
@@ -1,4 +1,4 @@
-// Copyright (C) 2004-2006 by Wired Team
+// Copyright (C) 2004-2007 by Wired Team
 // Under the GNU General Public License Version 2, June 1991
 
 // Copyright (C) 2004-2006 by Wired Team
@@ -165,18 +165,18 @@ class					SettingWindow : public wxDialog
   wxSizerFlags				BoxFlags;
 
  private:
-  /** Saves the selected index of a wxCheckListBox into a vector<long>.
+  /** Saves the selected index of a wxCheckListBox into a vector<int>.
    * \param from the wxCheckListBox where item are selected.
-   * \param to the vector<long> where indexes are saved.
+   * \param to the vector<int> where indexes are saved.
    */
   void					SaveChannels(wxCheckListBox* from, 
-						     std::vector<long>& to);
-  /** Selects into a wxCheckListBox indexes stored into a vector<long>.
+						     std::vector<int>& to);
+  /** Selects into a wxCheckListBox indexes stored into a vector<int>.
    * \param to the wxCheckListBox to fill.
-   * \param from the vector<long> containing the indexes.
+   * \param from the vector<int> containing the indexes.
    */ 
   void					LoadChannels(wxCheckListBox* to, 
-						     std::vector<long>& from);
+						     std::vector<int>& from);
 
 
   DECLARE_EVENT_TABLE()
